+++ gendarme/FSharpExamples/AvoidUnneededFieldInitializationRule.fs
@@ -1,0 +1,4 @@
+﻿namespace AvoidUnneededFieldInitialization
+
+type Prepare() =
+  member val LocalSource = false with get, set